fix(RHINENG-462-2): hide centos versions in os filter
Only inventory and tasks need to see centos versions in the operating system filter on the inventory table. All other apps will not be fetching centos systems, so they don't need them in the dropdown.

You can test this by running inventory with tasks and any other apps to make sure they show or are hidden.